done with laravel authentication and user privillage assignment

+++ app/Http/Controllers/PostsController.php
@@ -100,6 +100,11 @@ class PostsController extends Controller{
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return  redirect('/posts')->with('error','Unauthorized Page.');
+        }
+
         return  view('posts.edit')->with('post',$post);
     }
 
@@ -135,6 +140,10 @@ class PostsController extends Controller{
     public function destroy($id)
     {
         $post =Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','You are not authorized to delete this post');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success','Post deleted successfully');
